flaskapp/requirements.txt
Flask-Login==0.4.1
Flask-SQLAlchemy==2.4.1
stripe==2.37.2
